Muevo los hechizos a hechizos.cpp por comodidad.
