+++ archon.vcxproj
@@ -176,6 +176,7 @@
     <ClCompile Include="src\creditos.cpp" />
     <ClCompile Include="src\cursor.cpp" />
     <ClCompile Include="src\gallina.cpp" />
+    <ClCompile Include="src\hechizos.cpp" />
     <ClCompile Include="src\interaccion.cpp" />
     <ClCompile Include="src\ganador.cpp" />
     <ClCompile Include="src\juego.cpp" />
